+++ astro.config.mjs
@@ -1,0 +1,24 @@
+import { defineConfig } from 'astro/config';
+import react from '@astrojs/react';
+import tailwind from '@astrojs/tailwind';
+import path from 'path';
+import adminFsApiPlugin from './src/lib/admin-api-plugin.js';
+
+export default defineConfig({
+  integrations: [
+    react(),
+    tailwind({
+      applyBaseStyles: false,
+    }),
+  ],
+  vite: {
+    plugins: [
+      adminFsApiPlugin()
+    ],
+    resolve: {
+      alias: {
+        '@': path.resolve(process.cwd(), './src'),
+      },
+    },
+  },
+});